deleting blog requires being its creator

+++ part4/controllers/blogs.js
@@ -36,6 +36,20 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+
+  const blog = await Blog.findById(req.params.id)
+
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (!(blog.user.toString() === user.toString())) {
+    return res.status(401).json({ error: 'cannot delete other users blogs' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
